Change usage of parlay::random_number_generator to [] syntax
We intend to add unit testing to the SkipList data structure,
which requires a fixed seed to produce deterministic results. Changing
time(0) to an abitrary constant and then using [] rng syntax provides
this behaviour.

We may consider restoring the time(0) seed later (e.g., as a user input)
once the unit testing is in place.